# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Manage PDF documents",
        "description": "Simple API for manage PDF files",
        "version": "1.0.0"
    },
    "paths": {
        "/api/columns/": {
            "get": {
                "summary": "List columns",
                "description": "Returns a list of all colunms registered",
                "operationId": "list_columns",
                "responses": {
                    "200": {
                        "description": "Showing all Column types"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/columns": {
            "post": {
                "summary": "create a column",
                "description": "create a column",
                "operationId": "new_column",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ColumnRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Column type created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ColumnSchema"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "server error, try again"
                    }
                }
            }
        },
        "/api/columns/{id}": {
            "get": {
                "summary": "Get columns",
                "description": "Returns information about column by id",
                "operationId": "get_column_by_id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Column type retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ColumnSchema"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "server error, try again"
                    }
                }
            },
            "put": {
                "summary": "update a column",
                "description": "update a column",
                "operationId": "update_column",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ColumnRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Column type updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ColumnSchema"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "server error, try again"
                    }
                }
            },
            "delete": {
                "summary": "delete a column",
                "description": "delete a column",
                "operationId": "delete_column",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ColumnRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Column type deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ColumnSchema"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "server error, try again"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "ColumnRequest": {
                "properties": {
                    "id": {
                        "description": "id",
                        "type": "integer",
                        "example": "1",
                        "nullable": "false"
                    },
                    "name": {
                        "description": "name",
                        "type": "string",
                        "example": "string",
                        "nullable": "false"
                    },
                    "type_id": {
                        "description": "Type id",
                        "type": "integer",
                        "example": "1",
                        "nullable": "false"
                    }
                },
                "type": "object"
            },
...
            "ColumnSchema": {
                "properties": {
                    "id": {
                        "description": "id",
                        "type": "integer",
                        "example": "1",
                        "nullable": "false"
                    },
                    "name": {
                        "description": "name",
                        "type": "string",
                        "example": "string",
                        "nullable": "false"
                    },
                    "type_id": {
                        "description": "Type id",
                        "type": "integer",
                        "example": "1",
                        "nullable": "false"
                    }
                },
                "type": "object"
            },
...
        }
    }
}